Paaso 4 de aeropuerto

+++ GestionAeroportuari/src/AeropuertoPublico.java
@@ -1,16 +1,16 @@
 public class AeropuertoPublico extends Aeropuerto{
 
-    private double dineroNacion;
+    private float dineroNacion;
 
     public AeropuertoPublico(String nombre, String ciudad, String pais) {
         super(nombre, ciudad, pais);
     }
-    public AeropuertoPublico(String nombre, String ciudad, String pais, Company[] listCompanies, double dineroNacion) {
+    public AeropuertoPublico(String nombre, String ciudad, String pais, Company[] listCompanies, float dineroNacion) {
         super(nombre, ciudad, pais, listCompanies);
         this.dineroNacion = dineroNacion;
     }
 
-    public  AeropuertoPublico(String nombre, String ciudad, String pais, double dineroNacion){
+    public  AeropuertoPublico(String nombre, String ciudad, String pais, float dineroNacion){
         super(nombre, ciudad, pais);
         this.dineroNacion = dineroNacion;
     }
